+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Remove `float delta` parameter from `EngineService.update` method
 - `Family` introspection methods `requires` and `excludes`
 - Wired `List<T>`s are now resolved to the list of services that subclass `T`
+- Added optimized `EntitySetView.forEach(IntConsumer)` method and a few convenience methods
+- Change `FamilyWatcherSystem` to return `EntitySetView` instead of `Mask`
 
 # Version 0.2.5 (released 2020-01-21)
 - `EngineService`s are now automatically wired and initialized